+++ test/unit/Resources/TransactionTypes/ChargeTest.php
@@ -78,18 +78,16 @@ class ChargeTest extends BasePaymentTest
     }
 
     /**
-     * Setting recurrence type without a payment type Should raise Exception.
+     * Setting recurrence type without a payment type does not raise exception.
      *
      * @test
      */
-    public function SetRecurrenceTypeShouldRaiseExceptionWOPaymentType()
+    public function recurrenceTypeCanBeSetWithoutTypeParameter()
     {
         $charge = new Charge();
 
         $this->assertEmpty($charge->getAdditionalTransactionData());
         $this->assertEmpty($charge->getRecurrenceType());
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Payment type can not be determined. Set it first or provide it via parameter $paymentType.');
 
         $charge->setRecurrenceType(RecurrenceTypes::ONE_CLICK);
     }