+++ CHANGELOG.md
@@ -11,8 +11,10 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 *   Enable PHP version 8.2 in composer.json.
 *   Add paypage property to `Payment` class.
+*   Add class `\UnzerSDK\Resources\EmbeddedResources\CardTransactionData`.
+    Set as "card" field of additionalTransactionData    
 *   Add constants `\UnzerSDK\Constants\LiabilityShiftIndicator` for valid liability shift indicator values, relevant for card payment.
-    *   Charge and authorize transactions with card can contain that indicator in Api Response. (Accesed like:`$transaction->getAdditionalTransactionData()->card->liability`)
+    *   Charge and authorize transactions with card can contain that indicator in Api Response. (`CardTransactionData::$liability`)
 
 ### Changed
 
@@ -20,6 +22,9 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 *   Add Type declaration to methods where applicable.
 *   Enable `Unzer::fetchPaymentRefund` and `Unzer::fetchPaymentReversal` to use resource ID as parameter.
 *   Raise minimum PHP version to 7.4 in composer.json
+*   Handling of additionalTransactionData:
+    *   `additionalTransactionData.card` from API responses will be mapped on `\UnzerSDK\Resources\EmbeddedResources\CardTransactionData` now.
+    *   Remove `paymentType` parameter of `\UnzerSDK\Traits\HasRecurrenceType::setRecurrenceType`, it will always be set for `card` field of `additionalTransactionData`.
 
 ## [1.2.3.0](https://github.com/unzerdev/php-sdk/compare/1.2.2.0..1.2.3.0)